+++ pkg/manager/member/component_member_manager.go
@@ -17,14 +17,11 @@ import (
 	"fmt"
 	"path"
 	"path/filepath"
-	"reflect"
 	"regexp"
 	"strconv"
 	"strings"
 
-	"github.com/Masterminds/semver"
 	"github.com/pingcap/advanced-statefulset/client/apis/apps/v1/helper"
-	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
 	"github.com/pingcap/tidb-operator/pkg/label"
@@ -36,9 +33,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/klog"
-	podutil "k8s.io/kubernetes/pkg/api/v1/pod"
 	"k8s.io/utils/pointer"
 )
 
@@ -50,6 +45,8 @@ type ComponentContext struct {
 	component    string
 }
 
+// Sync Statefulsets
+// ---
 func ComponentSyncStatefulSetForTidbCluster(context *ComponentContext) error {
 	tc := context.tc
 	dependencies := context.dependencies
@@ -272,170 +269,9 @@ func ComponentSyncTidbClusterStatus(context *ComponentContext, set *apps.Statefu
 	return nil
 }
 
-func ComponentClusterVersionGreaterThanOrEqualTo4(version string) (bool, error) {
-	v, err := semver.NewVersion(version)
-	if err != nil {
-		return true, err
-	}
-
-	return v.Major() >= 4, nil
-}
-
-func collectUnjoinedPDMembers(context *ComponentContext, set *apps.StatefulSet, pdStatus map[string]v1alpha1.PDMember) error {
-	tc := context.tc
-	dependencies := context.dependencies
-
-	podSelector, podSelectErr := metav1.LabelSelectorAsSelector(set.Spec.Selector)
-	if podSelectErr != nil {
-		return podSelectErr
-	}
-	pods, podErr := dependencies.PodLister.Pods(tc.Namespace).List(podSelector)
-	if podErr != nil {
-		return fmt.Errorf("collectUnjoinedMembers: failed to list pods for cluster %s/%s, selector %s, error %v", tc.GetNamespace(), tc.GetName(), set.Spec.Selector, podErr)
-	}
-	for _, pod := range pods {
-		var joined = false
-		for pdName := range pdStatus {
-			ordinal, err := util.GetOrdinalFromPodName(pod.Name)
-			if err != nil {
-				return fmt.Errorf("unexpected pod name %q: %v", pod.Name, err)
-			}
-			if strings.EqualFold(PdName(tc.Name, ordinal, tc.Namespace, tc.Spec.ClusterDomain), pdName) {
-				joined = true
-				break
-			}
-		}
-		if !joined {
-			if tc.Status.PD.UnjoinedMembers == nil {
-				tc.Status.PD.UnjoinedMembers = map[string]v1alpha1.UnjoinedMember{}
-			}
-			ordinal, err := util.GetOrdinalFromPodName(pod.Name)
-			if err != nil {
-				return err
-			}
-			pvcName := ordinalPVCName(v1alpha1.PDMemberType, controller.PDMemberName(tc.Name), ordinal)
-			pvc, err := dependencies.PVCLister.PersistentVolumeClaims(tc.Namespace).Get(pvcName)
-			if err != nil {
-				return fmt.Errorf("collectUnjoinedMembers: failed to get pvc %s of cluster %s/%s, error %v", pvcName, tc.GetNamespace(), tc.GetName(), err)
-			}
-			tc.Status.PD.UnjoinedMembers[pod.Name] = v1alpha1.UnjoinedMember{
-				PodName:   pod.Name,
-				PVCUID:    pvc.UID,
-				CreatedAt: metav1.Now(),
-			}
-		} else {
-			if tc.Status.PD.UnjoinedMembers != nil {
-				delete(tc.Status.PD.UnjoinedMembers, pod.Name)
-			}
-		}
-	}
-	return nil
-}
-
-// ComponentShouldRecover checks whether we should perform recovery operation.
-func ComponentShouldRecover(context *ComponentContext) bool {
-	tc := context.tc
-	component := context.component
-	dependencies := context.dependencies
-
-	var pdMembers map[string]v1alpha1.PDMember
-	var stores map[string]v1alpha1.TiKVStore
-	var tidbMembers map[string]v1alpha1.TiDBMember
-	var pdFailureMembers map[string]v1alpha1.PDFailureMember
-	var failureStores map[string]v1alpha1.TiKVFailureStore
-	var tidbFailureMembers map[string]v1alpha1.TiDBFailureMember
-	var ordinals sets.Int32
-	var podPrefix string
-
-	switch component {
-	case label.TiKVLabelVal:
-		stores = tc.Status.TiKV.Stores
-		failureStores = tc.Status.TiKV.FailureStores
-		if failureStores == nil {
-			return false
-		}
-		ordinals = tc.TiKVStsDesiredOrdinals(true)
-		podPrefix = controller.TiKVMemberName(tc.Name)
-	case label.TiFlashLabelVal:
-		stores = tc.Status.TiFlash.Stores
-		failureStores = tc.Status.TiFlash.FailureStores
-		if failureStores == nil {
-			return false
-		}
-		ordinals = tc.TiFlashStsDesiredOrdinals(true)
-		podPrefix = controller.TiFlashMemberName(tc.Name)
-	case label.PDLabelVal:
-		pdMembers = tc.Status.PD.Members
-		pdFailureMembers = tc.Status.PD.FailureMembers
-		if pdFailureMembers == nil {
-			return false
-		}
-		ordinals = tc.PDStsDesiredOrdinals(true)
-		podPrefix = controller.PDMemberName(tc.Name)
-	case label.TiDBLabelVal:
-		tidbMembers = tc.Status.TiDB.Members
-		tidbFailureMembers = tc.Status.TiDB.FailureMembers
-		if tidbFailureMembers == nil {
-			return false
-		}
-		ordinals = tc.TiDBStsDesiredOrdinals(true)
-		podPrefix = controller.TiDBMemberName(tc.Name)
-	default:
-		klog.Warningf("Unexpected component %s for %s/%s in shouldRecover", component, tc.Namespace, tc.Name)
-		return false
-	}
-
-	// If all desired replicas (excluding failover pods) of tidb cluster are
-	// healthy, we can perform our failover recovery operation.
-	// Note that failover pods may fail (e.g. lack of resources) and we don't care
-	// about them because we're going to delete them.
-	for ordinal := range ordinals {
-		name := fmt.Sprintf("%s-%d", podPrefix, ordinal)
-		pod, err := dependencies.PodLister.Pods(tc.Namespace).Get(name)
-		if err != nil {
-			klog.Errorf("pod %s/%s does not exist: %v", tc.Namespace, name, err)
-			return false
-		}
-		if !podutil.IsPodReady(pod) {
-			return false
-		}
-		var exist bool
-		switch component {
-		case label.TiKVLabelVal, label.TiFlashLabelVal:
-			for _, v := range stores {
-				if v.PodName == pod.Name {
-					exist = true
-					if v.State != v1alpha1.TiKVStateUp {
-						return false
-					}
-				}
-			}
-		case label.PDLabelVal:
-			for pdName, pdMember := range pdMembers {
-				if strings.Split(pdName, ".")[0] == pod.Name {
-					if !pdMember.Health {
-						return false
-					}
-					exist = true
-					break
-				}
-			}
-		case label.TiDBLabelVal:
-			status, ok := tidbMembers[pod.Name]
-			if !ok || !status.Health {
-				return false
-			}
-		}
-
-		if !exist && component != label.TiDBLabelVal {
-			return false
-		}
-	}
-	return true
-}
-
+// Sync service
+// ---
 // sync service
-
 func componentSyncGeneralServiceForTidbCluster(context *ComponentContext, isHeadless bool) error {
 	tc := context.tc
 	dependencies := context.dependencies
@@ -452,9 +288,9 @@ func componentSyncGeneralServiceForTidbCluster(context *ComponentContext, isHead
 
 	var newSvc *corev1.Service
 	if isHeadless {
-		newSvc = ComponentGetNewServiceForTidbCluster(context)
-	} else {
 		newSvc = ComponentGetNewHeadlessServiceForTidbCluster(context)
+	} else {
+		newSvc = ComponentGetNewServiceForTidbCluster(context)
 	}
 
 	oldSvcTmp, err := dependencies.ServiceLister.Services(ns).Get(componentMemberName)
@@ -575,7 +411,7 @@ func componentGetNewGeneralServiceForTidbCluster(context *ComponentContext, isHe
 			svcName := controller.PDMemberName(tcName)
 			svcSelector := label.New().Instance(instanceName).PD()
 			svcLabels := svcSelector.Copy().UsedByEndUser().Labels()
-			componentService := &corev1.Service{
+			componentService = &corev1.Service{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:            svcName,
 					Namespace:       ns,
@@ -614,6 +450,7 @@ func componentGetNewGeneralServiceForTidbCluster(context *ComponentContext, isHe
 					componentService.Spec.Ports[0].Name = *svcSpec.PortName
 				}
 			}
+			fmt.Println(componentService.Name)
 		case label.TiKVLabelVal:
 			svcConfig := SvcConfig{
 				Name:       "peer",
@@ -632,7 +469,7 @@ func componentGetNewGeneralServiceForTidbCluster(context *ComponentContext, isHe
 				svcLabel = svcLabel.UsedByEndUser()
 			}
 
-			componentService := &corev1.Service{
+			componentService = &corev1.Service{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:            svcName,
 					Namespace:       ns,
@@ -690,7 +527,7 @@ func componentGetNewGeneralServiceForTidbCluster(context *ComponentContext, isHe
 				})
 			}
 
-			componentService := &corev1.Service{
+			componentService = &corev1.Service{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:            svcName,
 					Namespace:       ns,
@@ -720,6 +557,7 @@ func componentGetNewGeneralServiceForTidbCluster(context *ComponentContext, isHe
 			}
 		}
 	}
+
 	return componentService
 }
 
@@ -733,62 +571,8 @@ func ComponentGetNewHeadlessServiceForTidbCluster(context *ComponentContext) *co
 	return componentGetNewGeneralServiceForTidbCluster(context, isHeadless)
 }
 
-func ComponentStatefulSetIsUpgrading(set *apps.StatefulSet, context *ComponentContext) (bool, error) {
-	tc := context.tc
-	dependencies := context.dependencies
-
-	if statefulSetIsUpgrading(set) {
-		return true, nil
-	}
-	instanceName := tc.GetInstanceName()
-
-	componentLabel := getComponentLabel(context, instanceName)
-	selector, err := componentLabel.Selector()
-
-	if err != nil {
-		return false, err
-	}
-	componentPods, err := dependencies.PodLister.Pods(tc.GetNamespace()).List(selector)
-	if err != nil {
-		return false, fmt.Errorf("StatefulSetIsUpgrading: failed to list pods for cluster %s/%s, selector %s, error: %v", tc.GetNamespace(), instanceName, selector, err)
-	}
-
-	for _, pod := range componentPods {
-		revisionHash, exist := pod.Labels[apps.ControllerRevisionHashLabelKey]
-		if !exist {
-			return false, nil
-		}
-
-		componentUpdateRevision := getComponentUpdataRevision(context)
-
-		if revisionHash != componentUpdateRevision {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-func ComponentGetNewSetForTidbCluster(context *ComponentContext, cm *corev1.ConfigMap) (*apps.StatefulSet, error) {
-	component := context.component
-
-	switch component {
-	case label.PDLabelVal:
-		return componentPDGetNewSetForTiDBCluster(context, cm)
-	case label.TiKVLabelVal:
-		return componentTiKVGetNewSetForTiDBCluster(context, cm)
-	case label.TiFlashLabelVal:
-		return componentTiFlashGetNewSetForTiDBCluster(context, cm)
-	case label.TiDBLabelVal:
-		return componentTiDBGetNewSetForTiDBCluster(context, cm)
-	case label.TiCDCLabelVal:
-		return componentTiCDCGetNewSetForTiDBCluster(context, cm)
-	case label.PumpLabelVal:
-		return componentPumpGetNewSetForTiDBCluster(context, cm)
-	}
-
-	return &apps.StatefulSet{}, nil
-}
-
+// Sync ConfigMap
+// ---
 // ComponentSyncConfigMap syncs the configmap
 func ComponentSyncConfigMap(context *ComponentContext, set *apps.StatefulSet) (*corev1.ConfigMap, error) {
 	tc := context.tc
@@ -1398,135 +1182,27 @@ func syncComponentMembers(context *ComponentContext, set *apps.StatefulSet) erro
 	return nil
 }
 
-func getTiFlashStore(store *pdapi.StoreInfo) *v1alpha1.TiKVStore {
-	if store.Store == nil || store.Status == nil {
-		return nil
-	}
-	storeID := fmt.Sprintf("%d", store.Store.GetId())
-	ip := strings.Split(store.Store.GetAddress(), ":")[0]
-	podName := strings.Split(ip, ".")[0]
+// Sync Statefulsets
+// ---
+func ComponentGetNewSetForTidbCluster(context *ComponentContext, cm *corev1.ConfigMap) (*apps.StatefulSet, error) {
+	component := context.component
 
-	return &v1alpha1.TiKVStore{
-		ID:                storeID,
-		PodName:           podName,
-		IP:                ip,
-		LeaderCount:       int32(store.Status.LeaderCount),
-		State:             store.Store.StateName,
-		LastHeartbeatTime: metav1.Time{Time: store.Status.LastHeartbeatTS},
-	}
-}
-
-func setStoreLabelsForTiKV(context *ComponentContext) (int, error) {
-	tc := context.tc
-	dependencies := context.dependencies
-
-	ns := tc.GetNamespace()
-	// for unit test
-	setCount := 0
-
-	if !tc.TiKVBootStrapped() {
-		klog.Infof("TiKV of Cluster %s/%s is not bootstrapped yet, no need to set store labels", tc.Namespace, tc.Name)
-		return setCount, nil
+	switch component {
+	case label.PDLabelVal:
+		return componentPDGetNewSetForTiDBCluster(context, cm)
+	case label.TiKVLabelVal:
+		return componentTiKVGetNewSetForTiDBCluster(context, cm)
+	case label.TiFlashLabelVal:
+		return componentTiFlashGetNewSetForTiDBCluster(context, cm)
+	case label.TiDBLabelVal:
+		return componentTiDBGetNewSetForTiDBCluster(context, cm)
+	case label.TiCDCLabelVal:
+		return componentTiCDCGetNewSetForTiDBCluster(context, cm)
+	case label.PumpLabelVal:
+		return componentPumpGetNewSetForTiDBCluster(context, cm)
 	}
 
-	pdCli := controller.GetPDClient(dependencies.PDControl, tc)
-	storesInfo, err := pdCli.GetStores()
-	if err != nil {
-		return setCount, err
-	}
-
-	config, err := pdCli.GetConfig()
-	if err != nil {
-		return setCount, err
-	}
-
-	locationLabels := []string(config.Replication.LocationLabels)
-	if locationLabels == nil {
-		return setCount, nil
-	}
-
-	pattern, err := regexp.Compile(fmt.Sprintf(tikvStoreLimitPattern, tc.Name, tc.Name, tc.Namespace, controller.FormatClusterDomainForRegex(tc.Spec.ClusterDomain)))
-	if err != nil {
-		return -1, err
-	}
-	for _, store := range storesInfo.Stores {
-		// In theory, the external tikv can join the cluster, and the operator would only manage the internal tikv.
-		// So we check the store owner to make sure it.
-		if store.Store != nil && !pattern.Match([]byte(store.Store.Address)) {
-			continue
-		}
-		status := getTiKVStore(store)
-		if status == nil {
-			continue
-		}
-		podName := status.PodName
-
-		pod, err := dependencies.PodLister.Pods(ns).Get(podName)
-		if err != nil {
-			return setCount, fmt.Errorf("setStoreLabelsForTiKV: failed to get pods %s for cluster %s/%s, error: %s", podName, ns, tc.GetName(), err)
-		}
-
-		nodeName := pod.Spec.NodeName
-		ls, err := getNodeLabels(context, nodeName, locationLabels)
-		if err != nil || len(ls) == 0 {
-			klog.Warningf("node: [%s] has no node labels, skipping set store labels for Pod: [%s/%s]", nodeName, ns, podName)
-			continue
-		}
-
-		if !storeLabelsEqualNodeLabels(store.Store.Labels, ls) {
-			set, err := pdCli.SetStoreLabels(store.Store.Id, ls)
-			if err != nil {
-				msg := fmt.Sprintf("failed to set labels %v for store (id: %d, pod: %s/%s): %v ",
-					ls, store.Store.Id, ns, podName, err)
-				dependencies.Recorder.Event(tc, corev1.EventTypeWarning, FailedSetStoreLabels, msg)
-				continue
-			}
-			if set {
-				setCount++
-				klog.Infof("pod: [%s/%s] set labels: %v successfully", ns, podName, ls)
-			}
-		}
-	}
-
-	return setCount, nil
-}
-
-func getNodeLabels(context *ComponentContext, nodeName string, storeLabels []string) (map[string]string, error) {
-	dependencies := context.dependencies
-
-	node, err := dependencies.NodeLister.Get(nodeName)
-	if err != nil {
-		return nil, err
-	}
-	labels := map[string]string{}
-	ls := node.GetLabels()
-	for _, storeLabel := range storeLabels {
-		if value, found := ls[storeLabel]; found {
-			labels[storeLabel] = value
-			continue
-		}
-
-		// TODO after pd supports storeLabel containing slash character, these codes should be deleted
-		if storeLabel == "host" {
-			if host, found := ls[corev1.LabelHostname]; found {
-				labels[storeLabel] = host
-			}
-		}
-
-	}
-	return labels, nil
-}
-
-func storeLabelsEqualNodeLabels(storeLabels []*metapb.StoreLabel, nodeLabels map[string]string) bool {
-	ls := map[string]string{}
-	for _, label := range storeLabels {
-		key := label.GetKey()
-		if _, ok := nodeLabels[key]; ok {
-			val := label.GetValue()
-			ls[key] = val
-		}
-	}
-	return reflect.DeepEqual(ls, nodeLabels)
+	return &apps.StatefulSet{}, nil
 }
 
 func componentPDGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1.ConfigMap) (*apps.StatefulSet, error) {
@@ -2288,6 +1964,7 @@ func componentTiDBGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1.
 	tc := context.tc
 	ns := tc.GetNamespace()
 	tcName := tc.GetName()
+	setName := controller.TiDBMemberName(tcName)
 	headlessSvcName := controller.TiDBPeerMemberName(tcName)
 	baseTiDBSpec := tc.BaseTiDBSpec()
 	instanceName := tc.GetInstanceName()
@@ -2391,36 +2068,65 @@ func componentTiDBGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1.
 		podSecurityContext.Sysctls = []corev1.Sysctl{}
 	}
 
+	// handle StorageVolumes and AdditionalVolumeMounts in ComponentSpec
+	storageVolMounts, additionalPVCs := util.BuildStorageVolumeAndVolumeMount(tc.Spec.TiDB.StorageVolumes, tc.Spec.TiDB.StorageClassName, v1alpha1.TiDBMemberType)
+	volMounts = append(volMounts, storageVolMounts...)
+	volMounts = append(volMounts, tc.Spec.TiDB.AdditionalVolumeMounts...)
+
 	var containers []corev1.Container
+	slowLogFileEnvVal := ""
 	if tc.Spec.TiDB.ShouldSeparateSlowLog() {
 		// mount a shared volume and tail the slow log to STDOUT using a sidecar.
-		vols = append(vols, corev1.Volume{
-			Name: slowQueryLogVolumeName,
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{},
-			},
-		})
-		volMounts = append(volMounts, corev1.VolumeMount{Name: slowQueryLogVolumeName, MountPath: slowQueryLogDir})
+		var slowQueryLogVolumeMount corev1.VolumeMount
+		slowQueryLogVolumeName := tc.Spec.TiDB.SlowLogVolumeName
+		if slowQueryLogVolumeName == "" {
+			vols = append(vols, corev1.Volume{
+				Name: defaultSlowLogVolume,
+				VolumeSource: corev1.VolumeSource{
+					EmptyDir: &corev1.EmptyDirVolumeSource{},
+				},
+			})
+			slowQueryLogVolumeMount = corev1.VolumeMount{Name: defaultSlowLogVolume, MountPath: defaultSlowLogDir}
+			volMounts = append(volMounts, slowQueryLogVolumeMount)
+			slowLogFileEnvVal = defaultSlowLogFile
+		} else {
+			existVolume := false
+			for _, volMount := range storageVolMounts {
+				volMountName := fmt.Sprintf("%s-%s", v1alpha1.TiDBMemberType.String(), slowQueryLogVolumeName)
+				if volMount.Name == volMountName {
+					slowQueryLogVolumeMount = volMount
+					existVolume = true
+					break
+				}
+			}
+			if !existVolume {
+				for _, volMount := range tc.Spec.TiDB.AdditionalVolumeMounts {
+					if volMount.Name == slowQueryLogVolumeName {
+						slowQueryLogVolumeMount = volMount
+						existVolume = true
+						break
+					}
+				}
+			}
+			if !existVolume {
+				return nil, fmt.Errorf("Failed to get slowLogVolume %s for cluster %s/%s", slowQueryLogVolumeName, ns, tcName)
+			}
+			slowLogFileEnvVal = path.Join(slowQueryLogVolumeMount.MountPath, slowQueryLogVolumeName)
+		}
 		containers = append(containers, corev1.Container{
 			Name:            v1alpha1.SlowLogTailerMemberType.String(),
 			Image:           tc.HelperImage(),
 			ImagePullPolicy: tc.HelperImagePullPolicy(),
 			Resources:       controller.ContainerResource(tc.Spec.TiDB.GetSlowLogTailerSpec().ResourceRequirements),
-			VolumeMounts: []corev1.VolumeMount{
-				{Name: slowQueryLogVolumeName, MountPath: slowQueryLogDir},
-			},
+			VolumeMounts:    []corev1.VolumeMount{slowQueryLogVolumeMount},
 			Command: []string{
 				"sh",
 				"-c",
-				fmt.Sprintf("touch %s; tail -n0 -F %s;", slowQueryLogFile, slowQueryLogFile),
+				fmt.Sprintf("touch %s; tail -n0 -F %s;", slowLogFileEnvVal, slowLogFileEnvVal),
 			},
 		})
 	}
 
-	slowLogFileEnvVal := ""
-	if tc.Spec.TiDB.ShouldSeparateSlowLog() {
-		slowLogFileEnvVal = slowQueryLogFile
-	}
 	envs := []corev1.EnvVar{
 		{
 			Name:  "CLUSTER_NAME",
@@ -2460,11 +2166,6 @@ func componentTiDBGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1.
 		},
 	}
 
-	// handle StorageVolumes and AdditionalVolumeMounts in ComponentSpec
-	storageVolMounts, additionalPVCs := util.BuildStorageVolumeAndVolumeMount(tc.Spec.TiDB.StorageVolumes, tc.Spec.TiDB.StorageClassName, v1alpha1.TiDBMemberType)
-	volMounts = append(volMounts, storageVolMounts...)
-	volMounts = append(volMounts, tc.Spec.TiDB.AdditionalVolumeMounts...)
-
 	c := corev1.Container{
 		Name:            v1alpha1.TiDBMemberType.String(),
 		Image:           tc.TiDBImage(),
@@ -2500,7 +2201,7 @@ func componentTiDBGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1.
 	podSpec.Containers = append(containers, baseTiDBSpec.AdditionalContainers()...)
 	podSpec.Volumes = append(vols, baseTiDBSpec.AdditionalVolumes()...)
 	podSpec.SecurityContext = podSecurityContext
-	podSpec.InitContainers = initContainers
+	podSpec.InitContainers = append(initContainers, baseTiDBSpec.InitContainers()...)
 	podSpec.ServiceAccountName = tc.Spec.TiDB.ServiceAccount
 	if podSpec.ServiceAccountName == "" {
 		podSpec.ServiceAccountName = tc.Spec.ServiceAccount
@@ -2514,19 +2215,24 @@ func componentTiDBGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1.
 	podAnnotations := CombineAnnotations(controller.AnnProm(10080), baseTiDBSpec.Annotations())
 	stsAnnotations := getStsAnnotations(tc.Annotations, label.TiDBLabelVal)
 
+	deleteSlotsNumber, err := util.GetDeleteSlotsNumber(stsAnnotations)
+	if err != nil {
+		return nil, fmt.Errorf("get delete slots number of statefulset %s/%s failed, err:%v", ns, setName, err)
+	}
+
 	updateStrategy := apps.StatefulSetUpdateStrategy{}
 	if baseTiDBSpec.StatefulSetUpdateStrategy() == apps.OnDeleteStatefulSetStrategyType {
 		updateStrategy.Type = apps.OnDeleteStatefulSetStrategyType
 	} else {
 		updateStrategy.Type = apps.RollingUpdateStatefulSetStrategyType
 		updateStrategy.RollingUpdate = &apps.RollingUpdateStatefulSetStrategy{
-			Partition: pointer.Int32Ptr(tc.TiDBStsDesiredReplicas()),
+			Partition: pointer.Int32Ptr(tc.TiDBStsDesiredReplicas() + deleteSlotsNumber),
 		}
 	}
 
 	tidbSet := &apps.StatefulSet{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:            controller.TiDBMemberName(tcName),
+			Name:            setName,
 			Namespace:       ns,
 			Labels:          tidbLabel.Labels(),
 			Annotations:     stsAnnotations,
@@ -2850,4 +2556,41 @@ func componentTiCDCGetNewSetForTiDBCluster(context *ComponentContext, cm *corev1
 		},
 	}
 	return ticdcSts, nil
+}
+
+// Utils
+// ---
+func ComponentStatefulSetIsUpgrading(set *apps.StatefulSet, context *ComponentContext) (bool, error) {
+	tc := context.tc
+	dependencies := context.dependencies
+
+	if statefulSetIsUpgrading(set) {
+		return true, nil
+	}
+	instanceName := tc.GetInstanceName()
+
+	componentLabel := getComponentLabel(context, instanceName)
+	selector, err := componentLabel.Selector()
+
+	if err != nil {
+		return false, err
+	}
+	componentPods, err := dependencies.PodLister.Pods(tc.GetNamespace()).List(selector)
+	if err != nil {
+		return false, fmt.Errorf("StatefulSetIsUpgrading: failed to list pods for cluster %s/%s, selector %s, error: %v", tc.GetNamespace(), instanceName, selector, err)
+	}
+
+	for _, pod := range componentPods {
+		revisionHash, exist := pod.Labels[apps.ControllerRevisionHashLabelKey]
+		if !exist {
+			return false, nil
+		}
+
+		componentUpdateRevision := getComponentUpdataRevision(context)
+
+		if revisionHash != componentUpdateRevision {
+			return true, nil
+		}
+	}
+	return false, nil
 }
